Add HeatSourceBase class for common HeatSource implmentation
