+++ Assets/ModData.asset
@@ -31,7 +31,7 @@ MonoBehaviour:
   _steamTitle: 
   _steamVisibility: 0
   _versionMajor: 0
-  _versionMinor: 1051
+  _versionMinor: 1052
   _partCategories: []
   _parts: []
   _partStyleExtensions: []
